Added first iteration of MainMenu and GameMenu
Items now use names for comparisons.
Replaced item casting checks with string checks.
Bear Materials now correctly display.
Inventory items can now not be selected when the inventory is closed.
Clicking on an empty inventory space when an item is already selected now correctly displays and unselects the selected item.
GameManager now persists between Sessions

The Game now starts in the MainMenu scene, from which the player can start the game.
By pressing 'esc' in a level the GameMenu UI can be opened to exit to the MainMenu.

An implementation for Settings and Credits is required.
NewGame and Resume still have to be tested.
ISSUES WITH LOADING ARE STILL PRESENT!